+++ arestclient/Maven.py
@@ -1,16 +1,17 @@
 # Copyright 2012-2016 Jonathan Paugh and contributors
 # See COPYING for license details
-from agithub.base import API, ConnectionProperties, Client
+from arestclient.base import API, Client, ConnectionProperties
 
 
-class OpenWeatherMap(API):
+class Maven(API):
     """
-    Open Weather Map API
+    Maven Search API
     """
     def __init__(self, *args, **kwargs):
         props = ConnectionProperties(
-            api_url='api.openweathermap.org',
-            secure_http=False,
+            api_url='search.maven.org',
+            url_prefix='/solrsearch',
+            secure_http=True,
         )
         self.setClient(Client(*args, **kwargs))
         self.setConnectionProperties(props)
