+++ arestclient/OpenWeatherMap.py
@@ -1,17 +1,16 @@
 # Copyright 2012-2016 Jonathan Paugh and contributors
 # See COPYING for license details
-from agithub.base import API, Client, ConnectionProperties
+from arestclient.base import API, ConnectionProperties, Client
 
 
-class Maven(API):
+class OpenWeatherMap(API):
     """
-    Maven Search API
+    Open Weather Map API
     """
     def __init__(self, *args, **kwargs):
         props = ConnectionProperties(
-            api_url='search.maven.org',
-            url_prefix='/solrsearch',
-            secure_http=True,
+            api_url='api.openweathermap.org',
+            secure_http=False,
         )
         self.setClient(Client(*args, **kwargs))
         self.setConnectionProperties(props)
