+++ arestclient/__init__.py
@@ -1,5 +1,5 @@
 # Copyright 2012-2016 Jonathan Paugh and contributors
 # See COPYING for license details
-from agithub.base import VERSION, STR_VERSION
+from arestclient.base import VERSION, STR_VERSION
 
 __all__ = ["VERSION", "STR_VERSION"]
