+++ tox.ini
@@ -9,7 +9,7 @@ python =
 [testenv:flake8]
 basepython = python
 deps = flake8
-commands = flake8 agithub setup.py
+commands = flake8 arestclient setup.py
 
 [testenv]
 setenv =
